+++ meridian/model/media.py
@@ -50,7 +50,7 @@ def _roi_calibration_scaled_counterfactual(
     A tensor of scaled metric values with shape `(n_geos, n_times, n_channels)`
     where media values are set to zero during the calibration period.
   """
-  factors = tf.where(calibration_period, 0.0, 1.0)
+  factors = tf.cast(tf.where(calibration_period, 0.0, 1.0), dtype=tf.float64)
   return tf.einsum("gtm,tm->gtm", metric_scaled, factors)
 
 
@@ -101,10 +101,10 @@ def build_media_tensors(
     return MediaTensors()
 
   # Derive and set media tensors from media values in the input data.
-  media = tf.convert_to_tensor(input_data.media, dtype=tf.float32)
-  media_spend = tf.convert_to_tensor(input_data.media_spend, dtype=tf.float32)
+  media = tf.convert_to_tensor(input_data.media, dtype=tf.float64)
+  media_spend = tf.convert_to_tensor(input_data.media_spend, dtype=tf.float64)
   media_transformer = transformers.MediaTransformer(
-      media, tf.convert_to_tensor(input_data.population, dtype=tf.float32)
+      media, tf.convert_to_tensor(input_data.population, dtype=tf.float64)
   )
   media_scaled = media_transformer.forward(media)
   prior_type = model_spec.effective_media_prior_type
@@ -120,7 +120,7 @@ def build_media_tensors(
       input_data.aggregate_media_spend(
           calibration_period=calibration_period
       ),
-      dtype=tf.float32,
+      dtype=tf.float64,
   )
   # Set `prior_media_scaled_counterfactual` and `prior_denominator` depending on
   # the prior type.
@@ -141,7 +141,7 @@ def build_media_tensors(
     prior_denominator = aggregated_media_spend * (constants.MROI_FACTOR - 1.0)
   elif prior_type == constants.TREATMENT_PRIOR_TYPE_CONTRIBUTION:
     prior_media_scaled_counterfactual = None
-    total_outcome = tf.cast(input_data.get_total_outcome(), tf.float32)
+    total_outcome = tf.cast(input_data.get_total_outcome(), tf.float64)
     prior_denominator = tf.repeat(total_outcome, len(input_data.media_channel))
   elif prior_type == constants.TREATMENT_PRIOR_TYPE_COEFFICIENT:
     prior_media_scaled_counterfactual = None
@@ -185,11 +185,11 @@ def build_organic_media_tensors(
 
   # Derive and set media tensors from media values in the input data.
   organic_media = tf.convert_to_tensor(
-      input_data.organic_media, dtype=tf.float32
+      input_data.organic_media, dtype=tf.float64
   )
   organic_media_transformer = transformers.MediaTransformer(
       organic_media,
-      tf.convert_to_tensor(input_data.population, dtype=tf.float32),
+      tf.convert_to_tensor(input_data.population, dtype=tf.float64),
   )
   organic_media_scaled = organic_media_transformer.forward(organic_media)
 
@@ -248,11 +248,11 @@ def build_rf_tensors(
   if input_data.reach is None:
     return RfTensors()
 
-  reach = tf.convert_to_tensor(input_data.reach, dtype=tf.float32)
-  frequency = tf.convert_to_tensor(input_data.frequency, dtype=tf.float32)
-  rf_spend = tf.convert_to_tensor(input_data.rf_spend, dtype=tf.float32)
+  reach = tf.convert_to_tensor(input_data.reach, dtype=tf.float64)
+  frequency = tf.convert_to_tensor(input_data.frequency, dtype=tf.float64)
+  rf_spend = tf.convert_to_tensor(input_data.rf_spend, dtype=tf.float64)
   reach_transformer = transformers.MediaTransformer(
-      reach, tf.convert_to_tensor(input_data.population, dtype=tf.float32)
+      reach, tf.convert_to_tensor(input_data.population, dtype=tf.float64)
   )
   reach_scaled = reach_transformer.forward(reach)
   prior_type = model_spec.effective_rf_prior_type
@@ -261,7 +261,7 @@ def build_rf_tensors(
     calibration_period = tf.convert_to_tensor(calibration_period, dtype=tf.bool)
   aggregated_rf_spend = tf.convert_to_tensor(
       input_data.aggregate_rf_spend(calibration_period=calibration_period),
-      dtype=tf.float32,
+      dtype=tf.float64,
   )
   # Set `prior_reach_scaled_counterfactual` and `prior_denominator` depending on
   # the prior type.
@@ -281,7 +281,7 @@ def build_rf_tensors(
     prior_denominator = aggregated_rf_spend * (constants.MROI_FACTOR - 1.0)
   elif prior_type == constants.TREATMENT_PRIOR_TYPE_CONTRIBUTION:
     prior_reach_scaled_counterfactual = None
-    total_outcome = tf.cast(input_data.get_total_outcome(), tf.float32)
+    total_outcome = tf.cast(input_data.get_total_outcome(), tf.float64)
     prior_denominator = tf.repeat(total_outcome, len(input_data.rf_channel))
   elif prior_type == constants.TREATMENT_PRIOR_TYPE_COEFFICIENT:
     prior_reach_scaled_counterfactual = None
@@ -327,14 +327,14 @@ def build_organic_rf_tensors(
     return OrganicRfTensors()
 
   organic_reach = tf.convert_to_tensor(
-      input_data.organic_reach, dtype=tf.float32
+      input_data.organic_reach, dtype=tf.float64
   )
   organic_frequency = tf.convert_to_tensor(
-      input_data.organic_frequency, dtype=tf.float32
+      input_data.organic_frequency, dtype=tf.float64
   )
   organic_reach_transformer = transformers.MediaTransformer(
       organic_reach,
-      tf.convert_to_tensor(input_data.population, dtype=tf.float32),
+      tf.convert_to_tensor(input_data.population, dtype=tf.float64),
   )
   organic_reach_scaled = organic_reach_transformer.forward(organic_reach)
